move precip and mslp to derived channels
